// frontend/src/client/models/PasswordChange.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * 密码修改模式
 */
export type PasswordChange = {
    /**
     * 旧密码
     */
    old_password: string;
    /**
     * 新密码
     */
    new_password: string;
};